requirements.txt
langgraph
langchain
litellm

psycopg2-binary
pgvector
sqlalchemy
alembic

flask
streamlit
requests

pydantic
python-dotenv

openai
google-generativeai

tiktoken
numpy
pandas

uvicorn
fastapi
sentence-transformers
